fix(doctor): avoid helper runtime validation cycles
Move shared policy parsing into a standalone helper module so doctor
and consistency checks do not depend on each other, and reject
structurally invalid config payloads before reporting helper runtime
status.

This keeps doctor output reliable when config files are malformed and
addresses the follow-up Copilot review on PR #319.

Refs #305
